+++ Card/src/main/java/com/paypal/android/card/OrderRequest.kt
@@ -1,5 +1,0 @@
-package com.paypal.android.card
-
-import com.paypal.android.card.model.PurchaseUnit
-
-data class OrderRequest(val intent: OrderIntent, val purchaseUnits: List<PurchaseUnit>?)