+++ PayPalNativePayments/src/main/java/com/paypal/android/paypalnativepayments/PayPalNativeCheckoutError.kt
@@ -1,0 +1,11 @@
+package com.paypal.android.paypalnativepayments
+
+import com.paypal.android.corepayments.PayPalSDKError
+import com.paypal.checkout.error.ErrorInfo
+
+class PayPalNativeCheckoutError(
+    code: Int,
+    errorDescription: String,
+    correlationID: String? = null,
+    val errorInfo: ErrorInfo
+) : PayPalSDKError(code, errorDescription, correlationID)