+++ PayPalWebPayments/src/test/java/com/paypal/android/paypalwebpayments/PayPalWebCheckoutLifeCycleObserverUnitTest.kt
@@ -1,0 +1,18 @@
+package com.paypal.android.paypalwebpayments
+
+import io.mockk.mockk
+import io.mockk.verify
+import org.junit.Test
+
+class PayPalWebCheckoutLifeCycleObserverUnitTest {
+
+    @Test
+    fun `when resume is called, payPalClient handles browser switch result`() {
+        val payPalClient = mockk<PayPalWebCheckoutClient>(relaxed = true)
+
+        val lifeCycleObserver = PayPalWebCheckoutLifeCycleObserver(payPalClient)
+        lifeCycleObserver.resume()
+
+        verify { payPalClient.handleBrowserSwitchResult() }
+    }
+}